// src/app/models/district.type.ts
export type District =
    | "AVEIRO"
    | "BEJA"
    | "BRAGA"
    | "BRAGANÇA"
    | "CASTELO BRANCO"
    | "COIMBRA"
    | "ÉVORA"
    | "FARO"
    | "GUARDA"
    | "LEIRIA"
    | "LISBOA"
    | "PORTALEGRE"
    | "PORTO"
    | "SANTARÉM"
    | "SETÚBAL"
    | "VIANA DO CASTELO"
    | "VILA REAL"
    | "VISEU";